--- Nutrient_App/src/model/AuthValidation.java
@@ -1,8 +1,0 @@
-package src.model;
-
-public class AuthValidation {
-	
-//	public static isValidUsername(String s) {
-//		
-//	}
-}
