+++ Nutrient_App/src/server/DataServices/DBConfig.java
@@ -1,4 +1,4 @@
-package src.viewmodel.DataServices;
+package src.server.DataServices;
 import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStream;
