+++ Nutrient_App/src/server/DataServices/DBQueries.java
@@ -1,4 +1,4 @@
-package src.viewmodel.DataServices;
+package src.server.DataServices;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
