+++ Nutrient_App/src/server/authValidation/SignUpValidation.java
@@ -1,0 +1,5 @@
+package src.server.authValidation;
+
+public class SignUpValidation {
+
+}
