--- Nutrient_App/src/viewmodel/authValidation/LogInValidation.java
@@ -1,5 +1,0 @@
-package src.viewmodel.authValidation;
-
-public class LogInValidation {
-
-}
